Add files via upload
The Problem:
The Palmoria Group, a manufacturing company based in the Nigeria is embroiled in issues bordering on gender inequality in its 3 regions. Unfortunately, the media recently published in the news with the headline “Palmoria, the Manufacturing Patriarchy” This doesn’t look good for the owners of the business based on their ambition to scale
the business to other regions and even overseas. Cases like this can only spiral downwards revealing other issues like gender pay gap amongst other possible issues.

The Process:
The Data was cleaned using Microsoft Power BI and Microsoft Excel

The Results: 
Results are presented in the interactive Dashboard.